client: move Settings+EntitySourceDefinitions to services
These probably need to be moved together because the settings refer to
the name of the default entity source. (At least it was this way in
Repo. I haven’t actually tested it in Client.)

In the Scribunto tests, we need to change the way the settings are
managed, since in the service container they’ll be reset by the testing
framework between tests. Fortunately, the suite() override doesn’t seem
to be needed these days as far as I can tell, which means we can make
doMock() and unMock() non-static, which lets us call setService().

Bug: T275412
Change-Id: I3506ad80d57639f066c2b86358058b4432b6a666
Depends-On: I613f92b11df370f75e2012b58369147430d6cfae