Test alphabetical order of service wiring files
One wiring was already accidentally in the “wrong” order, which is fixed
here.

This uses TestingAccessWrapper to access the loaded wirings because I
don’t want to make the field in the base class protected instead of
private (normal service wiring tests shouldn’t need it), and I don’t
want this test to use the provideWiring() method either, since that’s a
data provider and returns a generator.

Bug: T275609
Change-Id: Ia6d9ec2b94cc1b7b66c4c4b479b54e45de4a5695