repo: move Logger to service container
Storing it in the service container feels slightly pointless, since the
LoggerFactory probably already keeps track of singleton loggers, but I
guess it’s still the right thing to do, both for consistency and so that
we can specify the service in extension-repo.json where needed.

Also extract some $services variables in some methods that were already
using the MediaWikiServices multiple times (implicitly or explicitly).

Bug: T275623
Change-Id: I750a5433e3d2ffc1de624f28cc4e28027e4215b9