Don’t pass local entity source name into RdfVocabulary
The parameter had already been unused since change Ifa0fe54220 (commit
89cbfe929b), which itself only removed a variable that had been unused
since change I4ae4198c1e (commit 1f23997e3f).

Because the class is instantiated a few times in other extensions, we
can’t just remove the constructor parameter and update all callers at
the same time. Instead, the constructor temporarily supports the old
calling convention as well, by detecting if a $localEntitySourceName
argument was passed in and shifting all other arguments in that case.
This will be removed soon.

EntitySource::getSourceName() gains a static return type declaration. As
far as I can tell, it’s already guaranteed to always return a string
(the constructor asserts that type and the field is never reassigned),
but I want to make absolutely sure, since the RdfVocabulary relies on
that type to detect whether it was called with a local entity source
name or not.

Bug: T275644
Change-Id: I2869c42d173c8e475c0899d81e2238079f7ab05e