repo: move TermsCollisionDetectorFactory to service container
In EditEntityFingerprintUniquenessIntegrationTest, SpecialNewItemTest
and SpecialNewPropertyTest we now need to reset services between API
calls, otherwise there’s a stuck NameTableStore somewhere with an
outdated copy of the wbt_type table which doesn’t have the “label” type,
and therefore the conflict isn’t detected. I suspect this can be removed
later, if we ensure that only one NameTableStore or DatabaseTypeIdsStore
is used throughout Wikibase, but we’re not there at the moment. (On the
bright side, we can remove ContentHandler::cleanupHandlersCache() calls,
since that method has been a deprecated no-op since MediaWiki 1.35.)

Bug: T276312
Change-Id: I14d689e74e2c01ca2de57a01390df0370af26e78