Test that services lists in extension JSON are ordered
We want the lists of services in each specification of the extension
JSON files to be ordered, just like we want the service wiring files to
be ordered. (Specifically, we want all MediaWiki services to be listed
before Wikibase services, though “Wikibase” is fairly late in
alphabetical order anyways.) Add tests to ensure that.

The tests are added to GlobalStateFactoryMethodsResourceTest, since that
already has convenient data providers for the relevant parts of the
extension JSON files. (Maybe those classes should be renamed at some
point, not sure.)

Change-Id: I628d2d9584023dfb53cc11c33c534aabd0b3a128