repo: move DatabaseTypeIdsStore to service container
The DatabaseTypeIdsStore contains an inner NameTableStore instance,
which caches a copy of the table (here, wbt_type) in an instance field.
By moving the store to the service container, we ensure that this cache
is shared across all users of the store; otherwise, the following
sequence of events is possible:

1. DatabaseTypeIdsStore instance A is created.
2. DatabaseTypeIdsStore instance B is created.
3. Instance A receives a call to resolveTypeIds() or lookupTypeIds(),
   causing the inner NameTableStore to load the table.
4. Instance B receives a similar call, and its inner NameTableStore also
   loads the table.
5. Instance A receives a call to acquireTypeIds(), with a type that has
   never been seen before. Its NameTableStore acquires the ID and writes
   it to the database, the WAN object cache, and its instance field.
6. Instance B receives a call to lookupTypeIds() with the type that was
   acquired in the previous step. Its inner NameTableStore still has the
   previous version of the table cached, and so it immediately throws a
   NameTableAccessException (no ID known for this type). It does not
   check the WAN object cache or the database, where that type could now
   be found.

I assume something along those lines happened in the tests that check
for label uniqueness (which involves loading/acquiring an ID for the
“label” term type), which is why change I14d689e74e had to add calls to
resetServices() to avoid breaking those tests. Now that all relevant
places share the same TypeIdsStore / NameTableStore, we can remove those
calls again, and the tests still pass.

This change actually introduces four services: the DatabaseTypeIdsStore,
which backs the other three but should not be used directly, and the
TypeIdsAcquirer, TypeIdsLookup and TypeIdsResolver, which are the
services that should be used when wiring up other services. We inject
all three of those services into the TermStoreWriterFactory, but e.g.
the TermsCollisionDetectorFactory only needs a TypeIdsLookup.

Bug: T276358
Change-Id: I78662c95b6c8e3608dd840eb08069043efd7df09