client: move Site to service container
The service wiring functions don’t have a useful __METHOD__, so replace
that for the log messages with a hard-coded string that’s hopefully not
too opaque for people looking at the logs.

The code to check the local IDs of the site is a bit buggy, but I’ll fix
that in a separate commit.

Bug: T276349
Change-Id: I9724724a635b808a7f1a8c88447216dc0d1dd738