repo: move LanguageFallbackChainFactory to service container
And also use it in a few places that had previously created their own
instance. In OutputPageBeforeHTMLHookHandler, inject it properly in the
constructor instead of getting it from global state in one method.

The two testHookFactoryShouldWork() tests are removed, since we already
test the factories of all registered hook handlers (including these two)
in GlobalStateFactoryMethodsResourceTest.

Bug: T276457
Change-Id: I26cbc8810b16faaf671eabe2c0ab2fe313cb7d45