Remove LanguageFallbackChainFactory from WikibaseServices
Both repo and client now have a version of this service in the MediaWiki
service container, so the WikibaseServices version is unused – except
for one place: in SingleEntitySourceServices. Fix that by injecting the
service there.

In GenericServicesTest and SingleEntitySourceServicesTest, I noticed
that the GenericServices constructor was being called with two extra
arguments, which had been removed from the constructor in change
Id1f424fd49 (commit 8ba4eae70a); fix that while we’re already touching
these files (though it’s not directly related to this change).

Bug: T276532
Change-Id: Ia6b59e8d9b6f5a6dd4e123bafef5ec060b367dcc