Protect against calling array_merge() without arguments
Site::getLocalIds() may return an empty array, and before PHP 7.4,
array_merge() may not be called with no arguments. Protect against the
resulting error by adding one empty array argument before the spread
argument.

A test is added for this case, where we mock getLocalIds() to return an
empty array. In this case, getLocalIds() will be called twice; I’ve
removed the $this->once() expectation on the method, since that seems
more sensible to me than asserting $this->exactly( 2 ). (It hardly seems
necessary to fail the test if the service wiring were changed to only
call getLocalIds() once and store the result in a variable.)

Bug: T276619
Change-Id: I3420b136f106838c453d4074b9d5b6c4a5793797