Inject more services into SpecialUnconnectedPages
Also inject a DBLoadBalancer instead of using wfGetDB(), and a
TitleFactory instead of using Title::newFromID().

For the DBLoadBalancer, we use the parent class’ getDBLoadBalancer()
method, but as far as I can tell from other special pages, we’re
supposed to take a load balancer in the constructor and pass it into
setDBLoadBalancer() – it’s not magically added by the environment.

Bug: T276933
Change-Id: I5385207156cbbb70ced683c75abeb09038bdfae4