repo: move SiteLinkBadgeChangeOpSerializationValidator to service container
I usually keep the service function headers (wiring and WikibaseRepo) in
one line, but in this case it just felt too long.

In SetSiteLinkTest, we need to override the settings differently so that
the validator service is reset and sees the new settings. (Calling
$this->resetServices() before changing the settings also works, but I
think calling setService() with a new settings instance is cleaner.)

Bug: T276580
Change-Id: If2b9380e1496c952e14507f63307a7e387a185d8