Fix LanguageFallbackChainFactory constructor
When the constructor is called with all three services provided, then it
should not try to access the default service container at all. This is
important when constructing a LanguageFallbackChainFactory in unit
tests, where that service container isn’t set up at all (otherwise we
get an error “premature access to service container”).

Bug: T277464
Change-Id: I5eb3d6be73ae1978ff37b37e8b64287b2fc9d4fb