Simplify DatabaseEntityTermsTableProvider logic drastically
This nested join in tables section is discouraged. Alias and prefixing
table names while being nice was added because the code needed to
self-join some tables but that got abandoned and it happens in the code
instead so there is no need to define alias.

This aliasing also breaks Postgres and any database that has a prefix
(including a lot of wikis in the wild).

Bug: T207226
Change-Id: Ic4f780bab43e8893d1501f629bb0b8530456d787