Add a sample rate for Lua function call tracking
Approximately 0.12% of MediaWiki wall-clock time is burnt on
incrementStatsKey() calls in Wikibase's mw.wikibase.lua and
mw.wikibase.entity.lua, which are expensive because they call into PHP.
This change adds a Lua wrapper for php.incrementStatKey that only
forwards a portion of calls to PHP, based on a configurable sample rate.
In PHP, the count is then scaled up by dividing it by the sample rate.
So if the sample rate is 0.1, only one in ten calls to
incrementStatsKey() will actually increment the counter, but the counter
will be incremented by 10.

Bug: T277817
Change-Id: I6b36c16473611e4972cad58b9b7713620a17aada