repo: remove EntityTerms factory method
It’s no longer needed – we can inject all the services directly into
the constructor, as long as we put the constructor parameters in the
right order.

Also, update the comments at the top of EntityTerms and PageTerms to
refer to “the services injected” instead of “the factory method”, since
one factory method is now gone and the other one will hopefully go away
as well soon.

Bug: T277774
Change-Id: I66c46ae6b8afcfeb15b202242f12dcba1e0274a2