client: move SiteGroup to service container
And change the fallback in case the siteGlobalID is not found in the
SiteLookup from true to “none”, as proposed in change I6bb3d2499e, so
that we at least always return a string and not sometimes a bool.

In GlobalStateFactoryMethodsResourceTest, we now need to configure the
siteGroup setting, otherwise the test fails: the default if no siteGroup
is configured uses the site store, which uses the database, causing the
test to fail. The only reason this didn’t fail before is that the
SiteGroup was already initialized very early in the test (as part of
addCoreDBData(), via a hook handler triggered when creating a test
page), before DB access was disallowed, and then it was cached in the
WikibaseClient instance (which was never reset). Now that the SiteGroup
is in the service container, it gets reset between tests, and therefore
its wiring code now runs when DB access is disallowed, causing the test
to fail. Avoiding DB lookups is exactly why the siteGroup setting was
introduced in the first place, so I think this is the right solution.

Bug: T277568
Change-Id: I81ec34a38c27ac2968f4fc12db5b1de1ad8e0f98
Co-Authored-By: Thiemo Kreuz <thiemo.kreuz@wikimedia.de>