repo: move Store to service container
And also inject the Settings into the Store while we’re at it. (The
SqlStore constructor is only called directly in very few places, so this
is not a very disruptive change.)

In DatabaseSchemaUpdater, we’re not allowed to inject the Store as a
service (the hook is run with the “noServices” option set, because “it
is called in a context where the global service locator is not
initialised”), but fortunately this class hasn’t used the Store since
change I18151cd2b4 (commit 4e56f7bcc5) anyways, so we can just get rid
of it along with the constructor and factory method.

In LinkTitles, inject not just the Store but also the SiteLookup,
because why not.

Bug: T277490
Change-Id: If695acdbe2b00c8be2bb243a2a4680fa1b142e1c