repo: move LocalRepoWikiPageMetaDataAccessor to service container
And inject the logger into WikiPageEntityMetaDataLookup. Since the class
is instantiated in WikibaseMediaInfo, temporarily make the default the
same as the old behavior (the Wikibase logger), but once MediaInfo
injects the right logger explicitly, I’ll change that to a NullLogger.

Note that the service wiring unit test still ends up relying on the
default MediaWikiServices instance, since it’s used in the constructor
of WikiPageEntityMetaDataLookup’s parent class, DBAccessBase. I don’t
yet know how best to fix that, but it’s also not the only test that
breaks when run as a true unit test (T277464), so I think that’s okay
for now.

Bug: T277489
Change-Id: Ic2293f9ccd888c82128f8a8ad75d9b5c189ba6a6