client: move Store to service container
WikibaseClient::overrideStore() is removed without replacement, since I
don’t think it can be reasonably reimplemented within WikibaseClient.
Callers should call $this->setService() instead (assuming they extend
MediaWikiIntegrationTest).

DataAccessSnakFormatterOutputFormatTest is moved from the unit tests to
the integration tests. It was very obviously an integration test before
(even containing a comment “this is an integration test”), but
apparently it happened to call no MediaWikiIntegrationTest methods
before. (It needs setService() now, see above.)

Integration tests that retrieved the current store, checked if it was a
mocked store, overrode it only if necessary, and then checked if the
store was overwritten successfully, are simplified to always override
the store and then assume this was successful, without checking again.

Bug: T277482
Change-Id: I6b80739641d55730d8e0a217d9f23aa9392bdf35
Depends-On: Ife404b9daab1b503426c7220019b245323b31c00