client: move RecentChangeFactory to service container
The call to CentralIdLookup::factoryNonLocal() isn’t very nice, but a
better alternative is not available yet (T265767). At least it doesn’t
break in unit tests.

The service wiring unit test is closely based on the old test in
WikibaseClientTest. In that class, getSiteLookup() is now no longer
needed and can be removed – none of the remaining tests require a
nonempty site lookup.

Bug: T277571
Change-Id: I7bf9f9364e07c70d5d972abefa2f309bb2b84d8c