repo: move SiteLinkConflictLookup to service container
The SiteLinkConflictLookup in the Store was only used when creating the
EntityConstraintProvider, so let’s create it there directly and remove
it from the Store interface, reducing the use of that interface as a
secondary service container. (Since it’s only used in one other service,
I see no need to expose it as a separate service, but I also wouldn’t
have a problem with that.)

Unfortunately this makes EntityConstraintProviderTest no longer a pure
unit test, since we now create a real SqlSiteLinkConflictLookup, which
makes the DBAccessBase constructor try to get the DBLoadBalancerFactory
from the global MediaWikiServices instance – see also the commit message
of change Ic2293f9ccd (commit 820a616f0e). We can fix that later.

Bug: T278829
Change-Id: I46213a220883aba8d43c540b58341ce42110db10