Tag JSON type of subentity statements
Previously, the statements ("claims") of subentities were emitted as
empty arrays rather than empty objects (if empty). This fix for that
mirrors change Ic388de009f (commit f2d0ed5499), which fixed a similar
issue where the datatype of a snak property was not being added to snaks
in subentity statements: we just look for "claims" keys anywhere two
levels below the entity (in the case of WikibaseLexemes, those levels
would be "forms" or "senses" followed by the sub-entity ID) and tag all
the arrays we find there.

Bug: T241422
Change-Id: Ia2080d8243ee54ca701f17e8cfaeb341705ff0da
Co-Authored-By: tzhelyazkova <tonina.zhelyazkova@wikimedia.de>