Add getEntityStore method to WikibaseApiTestCase
- Update calls to getEntityStore

Since there were so many API tests that use the Entity Store, this
change adds a method to the parent class of most of these tests to call
the service, and updates the reliant tests. Additionally, tests were
updated in test suites that do not extend WikibaseApiTestCase to call
the getEntityStore method statically.

Bug: T278828
Change-Id: I4d5e154b062247cd96cb5a6d761944bfc6e0f497