repo: move EntityLookup to service container
This is a weird service to migrate, since WikibaseRepo’s getEntityLookup
method has two optional parameters, a cache flag and a lookup mode. Our
approach is to move a version of the service with standard values for
those parameters to the service container, and any callers that need
nonstandard values should call getEntityLookup on a Store instead.

Because we can’t fix all such callers in this commit – some are in other
MediaWiki extensions – we still support the old calling convention in
WikibaseRepo::getEntityLookup() for a while: the method checks what kind
of arguments it was called with, and interprets them either as a service
container or as cache and lookup flags. WikibaseRepoTest gains several
tests, one per calling convention (once the compatibility code is
removed, those can all be deleted).

Currently, calling WikibaseRepo::getEntityLookup() with cache or lookup
mode arguments is soft-deprecated, but once all known callers in other
extensions are converted, we should make it hard-deprecated by emitting
deprecation warnings if the old calling conventions are used. (The code
for those warnings is already there, commented out, and the tests in
WikibaseRepoTest already anticipate those warnings as well, suppressing
them when testing the old calling convention, since I initially thought
we could hard-deprecate old-style calls immediately.)

Other calls to getEntityLookup() are not yet updated in this change, to
make it easier to review. They’ll be updated separately.

Bug: T278926
Change-Id: I1ead539ac6ef49673ffe6c0f357f0f7a2a373fc3