Hard-deprecate WikibaseRepo::getEntityLookup() with arguments
See change I1ead539ac6 for an explanation of the old and new calling
convention(s) of this method. All known users of the old calling
convention have been updated, so we can now hard-deprecate it; after one
or two weeks in production (assuming no other callers are discovered via
deprecation warnings in logstash), we can then remove support for it
completely.

Bug: T278926
Change-Id: I8d579908a198e17d9c69549253c20dfb497391b3
Depends-On: I164531f597f447aaf3b3021830cbff1690ae64c2
Depends-On: I6916e5b2d58aa328a496f85161b9f152be7e9b15