Fix MediaWiki version in deprecation warning again
In change I91a86b5e32 (commit be14a0f1cf), I changed this from 1.35 to
1.37, because I saw that the Phabricator task had been tagged for
inclusion in 1.37-wmf.1, rather than 1.36-wmf.something. But the REL1_36
branch hadn’t been cut yet, and it was only now cut from current master,
so the deprecation is actually on the REL1_36 branch as well.

I hope that we’ll completely remove this backwards compatibility code
soon (T279563), and that we’ll also backport this removal to REL1_36
before MediaWiki 1.36 is released, but in case that doesn’t happen,
let’s make sure the version number is accurate.

Bug: T278926
Change-Id: I0749d3c5aaff2d8dd5f25194fb12a1cf3b0156e1