repo: get SiteLookup from MediaWikiServices
This changes the remaining callers of WikibaseRepo::getSiteLookup()
outside of the WikibaseRepo class to call the MediaWikiServices method
instead, and changes the WikibaseRepo method to private to ensure no one
else starts calling it again.

Callers inside WikibaseRepo are left alone because we’ll take care of
those sooner or later anyways, as we migrate the remaining services to
the service wiring file.

Bug: T279681
Change-Id: Icfe4078a0e12d75d2efa4c1392335902cebd1c68