repo: move EntityDataUriManager to service container
Unfortunately, the wiring test crashes as when run as a unit test, since
SpecialPage::getTitleFor() eventually uses the global service container,
and I haven’t found a way to avoid that (short of inlining the whole
method). I guess we’ll have to live with it for now.

In PurgeEntityDataJob::newFromGlobalState(), also fix the call to
WikibaseRepo::getEntityIdParser(), which was migrated to the service
container ages ago (change I851a0783d4, commit 15b69c350e).

Bug: T279913
Change-Id: Ib07baef1c8f1392fba0c435a39ff1111f9f04409