repo: move ItemHandler to service container
And also rename the accessor method from newItemHandler() to
getItemHandler() – I assume nothing actually relied on it returning a
new instance each time. (The old name is kept for now, since it’s used
in one other extension, WikibaseCirrusSearch.)

The tests for the transformLegacyFormatOnExport setting in
WikibaseRepoTest are removed without direct replacement – I think this
is covered by the wiring tests for the LegacyFormatDetectorCallback
service, and it doesn’t seem necessary to test specifically that the
ItemHandler class and wiring pass through that callback correctly.

Bug: T279939
Change-Id: I590f2dd177cfe093ecc71e6100dddc2f5bb6c5f5