repo: move DefaultValidatorBuilders to service container
There are practically no callers to migrate, because the method is
already static and the non-test callers (in the various data type wiring
files, including in other extensions) have no particular $services value
to inject.

The MediaWikiPageNameNormalizer() gets a TODO, since that class should
probably be modernized, but I’m not tackling that right now. (It’s also
used in the CachingCommonsMediaFileNameLookup service, by the way.)

Bug: T279529
Change-Id: I789517d39c2e395298bcb5cde762c5ca905de6ae