repo: move PropertyHandler to service container
And also rename the accessor method from newPropertyHandler() to
getPropertyHandler() – I assume nothing actually relied on it returning
a new instance each time. (The old name is kept for now, since it’s used
in one other extension, WikibaseCirrusSearch.)

The tests for the transformLegacyFormatOnExport setting in
WikibaseRepoTest are removed without direct replacement – I think this
is covered by the wiring tests for the LegacyFormatDetectorCallback
service, and it doesn’t seem necessary to test specifically that the
PropertyHandler class and wiring pass through that callback correctly.

Bug: T280143
Change-Id: Icbe0c568b033de019f90bbd9d62b1d8091d13964