Skip trying to format ids that don't exist
The result from the database can be cached and return an id that doesn't
exist anymore. That's why TitleFactory returns null and breaks
SpecialUnconnectedPages. Lots of other QueryPage implementations just
skip rendering of the row if the title is null. This is copying what has
been done in those classes.

Also writing some tests for it.

Bug: T279408
Change-Id: I0a0c9217d74c1f42e19f3b9b6d192ae33246622a