Move WikibaseRepo EditEntityFactory to service container
- Add getEditEntityFactory accessor method
- Update calls from newEditEntityFactory to getEditEntityFactory
- Update newEditEntityFactory to alias getEditEntityFactory and
  deprecate it

In this change we decided to deprecate the usage of any Request
context other than the default one from RequestContext::getMain().
Therefore, the $context parameter is no longer needed since - as was
confirmed by tests in earlier patches of this change, the only non
default contexts that were passed to this method were test contexts.

The deprecated newEditEntityFactory method will be removed after
all callers are updated to the getEditEntityFactory accesor

Bug: T280140
Change-Id: If9cda35b69b00e7db537da697e3a40da4ca040ef