Move EditFilterHookRunner to service container
- Rename newEditFilterHookRunner to getEditFilterHookRunner
- Update service wiring and tests to use the new service

This change deprecates the ability to pass non default
contexts for use with EditFilterHookRunner. Similarily to
what was done in If9cda35b, it was confirmed by CI tests
in earlier patches of this change that the only non default
contexts that were passed to this method were test contexts.

Bug: T280346
Change-Id: Id46581a3a71ccd71fc0e034a8ba1a284d5c27763