ruby tests: update gems for newer watir version
This sets HEADLESS_CAPTURE_PATH to nil because libav is not installed in
CI, which means that there will be no video recording. This is a hack
and ideally we would get libav installed on the container running the
tests.

The majority of necessary changes resulting from the major version
update of selenium-webdriver are simply switching `when_not_visible` to
`when_not_present`.

Bug: T280491
Change-Id: I2dc84c490e856a329af79d073d16a20aaa7f85e1