Make newEditEntity() accept context argument
This is the first part of a refactoring to ensure that a context is no
longer passed into the MediawikiEditFilterHookRunner at construction
time, but rather is passed through later. Here, we change the first
argument of MediawikiEditEntityFactory::newEditEntity() to a context, so
that this context can later be passed into MediawikiEditEntity and from
there into EditFilterHookRunner::run().

Because MediawikiEditEntityFactory::newEditEntity() is also called from
WikibaseLexeme, we still accept a user argument as well, and that is
also why we don’t yet pass the context into the MediawikiEditEntity and
beyond – that will be done later. For now, we only adjust the callers in
Wikibase to pass a context into newEditEntity() rather than a user.

In the majority of callers (especially the WikibaseLexeme ones), the
user that was previously passed into newEditEntity() came directly from
a context, and so it’s straightforward to replace the user with the
context. (Minor note: the parameter type is IContextSource, which is
implemented by API modules and other classes, so the call to
->getContext() might seem to be redundant; however, the object should
really be a MutableContext – MediawikiEditFilterHookRunner will log a
warning if it isn’t – and that’s what the ->getContext() call
accomplishes.) However, in two places in Wikibase, UpdateRepoJob and
populateWithRandomEntitiesAndTerms.php, the method is actually called
with a custom user. In these cases, we construct a DerivativeContext
based on the context and explicitly set its user to the one that we want
to make the edit with. I think this is an improvement anyways:
previously, the MediawikiEditFilterHookRunner would eventually end up
with both a context and a user (context from the constructor, user as a
method argument), but the context would have a different user!

Bug: T280365
Change-Id: I875584a1aa0a8a355c625a0e7626d064399848d9