Make EntityRedirectCreationInteractor accept context argument
This is another change in preparation for ensuring that a context is no
longer passed into the MediawikiEditFilterHookRunner at construction
time. It’s similar to change I875584a1aa, which changed the parameter
type of MediawikiEditEntityFactory::newEditEntity(); like that method,
the EntityRedirectCreationInteractor constructor is called from
WikibaseLexeme, and so we temporarily need to allow both context and
user arguments (and continue to pass only a user through), before we can
later change the signature to accept only a context, and use that
context later (in the case of EntityRedirectCreationInteractor, directly
passing it into EditFilterHookRunner::run()).

This means that WikibaseRepo::newItemRedirectCreationInteractor() now
uses its context argument again, rather than the user argument. Ideally,
it should use neither, and here too the context should be passed through
at a later point, not at service construction time; however, that will
have to be done separately.

In the two special page tests, there was already a hack to make the test
user available in newSpecialPage() (this hack underscores that we really
shouldn’t need a context or user at all to create the special page);
that hack is here expandedd to also make the request available, and then
use that to build a context. (Similar code to create a context from a
request was already present in other tests, and is here reused.)

Bug: T280365
Change-Id: Ibba3955a6efbe796bb11e7fa89fe7b9ae5aab02b