Lua: resolve redirects for `getSitelink`
This uses EntityRevisionLookup to find out the Item's potential redirect
target, which is not ideal becauses the revision is unused. T280771 was
created to look into a better solution for resolving redirects in a
follow-up.

If the item passed to `getSitelink` is a redirect, usage will be tracked
for both the redirect (all usage) and the redirect target (sitelink
aspect).

Bug: T157868
Bug: T280758
Change-Id: I6ee78c86b4937673b7f21f1889bde4678bf380bd