Make WikibaseClient constructor private
We don’t want anyone to be able to create an instance of WikibaseClient,
so add a private constructor to mask the default public one. (We could
even throw an exception, but I feel like that’s not worth it.)

I did a cursory codesearch for other private no-argument constructors,
but there doesn’t seem to be a standard pattern for their contents
(comment phrasing or exception message) that we should follow.

Bug: T280803
Change-Id: I186cae6d7a9817d38f9ce42a97dc82e0f8e458ab