Remove isRepoEnabled() checks in repo tests
The integration tests in repo/tests/phpunit/includes/ generally assume
that WikibaseRepo is enabled; I don’t think it’s useful to explicitly
check this condition in a few tests.

In the two tests touched in this change, it seems that the tests were
moved out of lib/ at some point, which might explain why the check used
to be there.

Change-Id: I9740f64d328189a2ae79ca723f016087468fc0c7