Inject ILBFactory into WikiPageEntityMetaDataLookup
We inject a load balancer factory, not a load balancer, since the
responsibility of getting the database name from the entity source was
previously inside the WikiPageEntityMetaDataLookup constructor, and I
think it makes sense to keep it this way. I could also live with the
alternative, though.

This makes the LocalRepoWikiPageMetaDataAccessor service wiring test a
proper unit test, since the WikiPageEntityMetaDataLookup no longer gets
its load balancer (factory) from the global service container. (In
SingleEntitySourceServices we still use the global service container for
the time being – this isn’t the first place where that happens.)

Calls to releaseConnection() are removed without replacement, since that
method is already an empty no-op in IDBAccessObject – explicitly
releasing a connection is unnecessary when using getConnectionRef(). And
while we’re at it, also get the DB_* constants from the ILoadBalancer
interface, and stop using the deprecated wfGetDB() function.

Because the WikiPageEntityMetaDataLookup constructor is called in
WikibaseMediaInfo, we temporarily allow the old calling convention as
well, where the fourth argument is a logger. This will be removed in
change I678c16dbb1.

Bug: T281116
Change-Id: I535fc3b7fcd9d7c39bf987e7f41c329d463f5701