bridge: Switch tests waiting for bridge to open from 10 -> 20s wait
TLDR of the main investigation ticket leading to the 10 -> 20s
change at T281122 is that sometimes the VMs that run our CI get
slow.
After looking at the most recent failures of this test case, a
20s timeout instead of a 10s timeout would have lead to them all
passing.

If our CI ever speeds up perhaps we can decrease this again?
But for now a longer timeout here and a more stable CI system
will be a big win for everyone! (IMO)

Bug: T281122
Bug: T277205
Bug: T276932
Change-Id: I61c0d8f3d4808c05e90d79d45681005b51ff9ba9