bridge: Run tests with window-size 300px again
There was apparently some issue of a bit too much test code being copied
when extracting the test on mobile from the one on desktop.

The mobile test is moved to its own test-suite to encapsulate tests on a
mobile viewport. The reason behind this is that it is apparently not
possible with webdriver to define setup/teardown methods on a
per-testsuite basis. This means that a method to reset the window to at
least desktop dimensions would have to be run after each test in the
entire Wikidata Bridge Selenium test setup, which seems excessive.

Bug: T280361
Change-Id: I6710ef0bd56b4e7613e64b025a9677684e21b5ac