Replace WikibaseRepoTest with tests for service accessors
This is effectively a complete rewrite of WikibaseRepoTest. For every
WikibaseRepo method (except the constructor), we now test:

* it’s public, static, named getSomething(), and has a return type;
* when called with a custom service container, it gets the right service
  from that container; and
* when called without arguments, it gets the right service from the
  default MediaWikiServices.

The old tests are almost completely removed, since they were for the
most part outdated and superseded by the individual service wiring tests
(which also disallow DB and HTTP access). The WikibaseServices test is
moved to GlobalStateFactoryMethodsResourceTest, since I think it makes
more sense in that class (eventually, of course, we’d like to get rid of
WikibaseServices entirely); the entitySourceBasedFederation flag is
removed, though, since Wikibase stopped supporting it a while ago.

createValue() is a small helper function to create some value of any PHP
type, primitive or not; as far as I could tell, there’s no such function
in PHPUnit, though I’d be happy to be proven wrong. (For non-primitive
types, we of course fall back to createMock() – we just need to take
care of primitive types, which createMock() doesn’t like.)

Bug: T281556
Change-Id: I400dec13f305b02b262f795f9f588c20afbf2f85