Resurrect WikibaseClientTest
This recreates WikibaseClientTest as a copy of the new WikibaseRepoTest,
testing only that the service accessor methods look the way they’re
supposed to and do the right thing. Similarly, a test for
WikibaseServices is added to GlobalStateFactoryMethodsResourceTest,
though this time we don’t change the settings. See the parent change for
more details.

Bug: T281556
Change-Id: I70f464f5bc2912755620af6b55d85cf8283b3fdb