Introduce MessageInLanguageProvider service
This new service is similar to MediaWiki’s MessageLocalizer, but takes
the language of the message as an argument, which means it doesn’t
depend on the user language, which allows us to add it to the service
container. We then use the service in OutputFormatSnakFormatterFactory,
which makes SummaryFormatterTest a pure unit test (previously, it ended
up calling wfMessage(), which would use the default service container –
now we can inject a mock MessageInLanguageProvider instead).

Bug: T281677
Change-Id: Ie8453cba156dcf4ef3056d7d86b0a2d9e3cd2c18