Introduce LanguageNameLookupFactory repo service
This will replace some uses of LanguageNameLookup, which is an ugly
service because it requires the user language. LanguageNameLookupFactory
allows us to still use the LanguageNameLookup class, but only get an
instance at a later time, not at service creation time.

The two methods, getForLanguage() and getForLanguageCode(), are modeled
after NumberLocalizerFactory and LocalizedTextProviderFactory.

The LanguageNameLookup service itself is reimplemented in terms of
LanguageNameLookupFactory.

Bug: T281712
Change-Id: I5022ba489234f58775603bdbdb1a90cb30863329