CreateClaimTest: Use a string Property for testing
This test was changed to use a commonsMedia Property in late 2012,
in acd8f421ce. Adding commonsMedia Snaks invokes secondary data
updates, that (try to) access the media storage, which is undesirable
in tests. Given that these side effects/ secondary data updates
didn't exist when the original change was made, it is safe to assume
that this was not done in an attempt to also test these.

Bug: T268890
Change-Id: I581c0deb1bd2611ad705cff0cbe917a4522deb0c