Allow failures of secondary CI on REL1_36
Until the ADR 0018 [0] is accepted, we probably do not want secondary CI to
fail if that particular check isn't working. At the same time it makes
sense to keep this check running in order to make it easier to spot
upcoming problems.

The `continue-on-error` syntax [1] used in this patch isn't exactly the
same as what `allow-failure` was on Travis-CI, but it seems close
enough.

We also have to add the condition that decides whether to send an email
on a failed CI-check as still gets executed even with
continue-on-error.

[0] https://github.com/wikimedia/Wikibase/blob/48324e796f04e067d0449031cac35b683062e5be/docs/adr/0018-wikibase-compatibility-with-mediawiki.md
[1] https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idcontinue-on-error

Change-Id: I496f2c9b4ce88e442be92881cc19ced8dcb70a9a