Extract RecentChangeSaveHookHandler from RepoHooks
This hook handler does a lot of things. Extracting it makes it more
testable and more easily digestible. Also, this is the only place where
RepoEntityChange::setMetadataFromRC is ever used and likely the only
place where it is useful. It will be moved into the hook handler in a
following patch.

Bug: T282697
Change-Id: Id36dc54780bce722eefee7f86935a7f23cc2a6b1