bridge: standardize timeouts in selenium tests
This now uses consistently 20s for waitFor* functions that (implicitly)
wait for API calls to return and 10s otherwise.

To not repeat `10000` all over the code, a custom configuration is used:
https://webdriver.io/docs/browserobject/#custom-configurations

This is also the timeout used for awaiting the license popup as there is
no API call involved with that popup either.

Util.waitForModuleState is a custom method with a hard-coded default
timeout of 2 seconds, so the timeout parameter is explicitly provided to
ensure we wait for up to 20 seconds.

The way waiting is done for error states is intended to follow these
lines of reasoning:
* Waiting for a permissions error requires API calls to succeed and
  therefore gets the default 20 seconds.
* Waiting for an error due to unavailable network doesn't require API
  calls to succeed and therefore only gets 10 seconds of wait time.

Bug: T281339
Change-Id: I3a5ca456891b44d57ab9d2395df6a7084dd8b8e1